# Add AttributeMap and use it in the Resource