--- a/internal/data/common.go
+++ b/internal/data/common.go
@@ -128,7 +128,161 @@ func (a AttributeValue) SetBool(v bool) {
 	a.orig.BoolValue = v
 }
 
+// AttributeKeyValue stores a key and AttributeValue pair.
+type AttributeKeyValue struct {
+	orig *otlpcommon.AttributeKeyValue
+}
+
+// NewAttributeKeyValue creates a new AttributeKeyValue with the given key.
+func NewAttributeKeyValue(k string) AttributeKeyValue {
+	return AttributeKeyValue{&otlpcommon.AttributeKeyValue{Key: k}}
+}
+
+// NewAttributeKeyValueString creates a new AttributeKeyValue with the given key and string value.
+func NewAttributeKeyValueString(k string, v string) AttributeKeyValue {
+	akv := AttributeKeyValue{&otlpcommon.AttributeKeyValue{Key: k}}
+	akv.Value().SetString(v)
+	return akv
+}
+
+// NewAttributeKeyValueInt creates a new AttributeKeyValue with the given key and int64 value.
+func NewAttributeKeyValueInt(k string, v int64) AttributeKeyValue {
+	akv := AttributeKeyValue{&otlpcommon.AttributeKeyValue{Key: k}}
+	akv.Value().SetInt(v)
+	return akv
+}
+
+// NewAttributeKeyValueDouble creates a new AttributeKeyValue with the given key and float64 value.
+func NewAttributeKeyValueDouble(k string, v float64) AttributeKeyValue {
+	akv := AttributeKeyValue{&otlpcommon.AttributeKeyValue{Key: k}}
+	akv.Value().SetDouble(v)
+	return akv
+}
+
+// NewAttributeKeyValueBool creates a new AttributeKeyValue with the given key and bool value.
+func NewAttributeKeyValueBool(k string, v bool) AttributeKeyValue {
+	akv := AttributeKeyValue{&otlpcommon.AttributeKeyValue{Key: k}}
+	akv.Value().SetBool(v)
+	return akv
+}
+
+// Key returns the key associated with this AttributeKeyValue.
+func (akv AttributeKeyValue) Key() string {
+	return akv.orig.Key
+}
+
+// Value returns the value associated with this AttributeKeyValue.
+func (akv AttributeKeyValue) Value() AttributeValue {
+	return AttributeValue{akv.orig}
+}
+
+// SetValue replaces the value associated with this AttributeKeyValue.
+func (akv AttributeKeyValue) SetValue(av AttributeValue) {
+	akv.orig.Type = av.orig.Type
+	akv.orig.StringValue = av.orig.StringValue
+	akv.orig.IntValue = av.orig.IntValue
+	akv.orig.DoubleValue = av.orig.DoubleValue
+	akv.orig.BoolValue = av.orig.BoolValue
+}
+
+// AttributeMap stores a map of attribute keys to values.
+type AttributeMap struct {
+	orig *[]*otlpcommon.AttributeKeyValue
+}
+
+// NewAttributeMap creates a new AttributeMap from the given map[string]AttributeValue.
+func NewAttributeMap(attrMap map[string]AttributeValue) AttributeMap {
+	origs := make([]otlpcommon.AttributeKeyValue, len(attrMap))
+	wrappers := make([]*otlpcommon.AttributeKeyValue, len(attrMap))
+
+	ix := 0
+	for k, v := range attrMap {
+		wrappers[ix] = &origs[ix]
+		wrappers[ix].Key = k
+		AttributeKeyValue{wrappers[ix]}.SetValue(v)
+		ix++
+	}
+
+	return AttributeMap{&wrappers}
+}
+
+// Get returns the AttributeKeyValue associated with the key and true,
+// otherwise an invalid instance of the AttributeKeyValue and false.
+func (am AttributeMap) Get(key string) (AttributeKeyValue, bool) {
+	for _, a := range *am.orig {
+		if a.Key == key {
+			return AttributeKeyValue{a}, true
+		}
+	}
+	return AttributeKeyValue{nil}, false
+}
+
+// Delete deletes the entry associated with the key and returns true if the key
+// was present in the map, otherwise returns false.
+func (am AttributeMap) Delete(key string) bool {
+	for i, a := range *am.orig {
+		if a.Key == key {
+			(*am.orig)[i] = (*am.orig)[len(*am.orig)-1]
+			*am.orig = (*am.orig)[:len(*am.orig)-1]
+			return true
+		}
+	}
+	return false
+}
+
+// Insert adds the AttributeKeyValue to the map when the key does not exist.
+// No action is applied to the map where the key already exists.
+func (am AttributeMap) Insert(akv AttributeKeyValue) {
+	if _, existing := am.Get(akv.Key()); !existing {
+		*am.orig = append(*am.orig, akv.orig)
+	}
+}
+
+// Update updates an existing AttributeKeyValue with a value.
+// No action is applied to the map where the key does not exist.
+func (am AttributeMap) Update(akv AttributeKeyValue) {
+	if av, existing := am.Get(akv.Key()); existing {
+		av.SetValue(akv.Value())
+	}
+}
+
+// Upsert performs the Insert or Insert action. The AttributeKeyValue is
+// insert to the map that did not originally have the key. The key/value is
+// updated to the map where the key already existed.
+func (am AttributeMap) Upsert(akv AttributeKeyValue) {
+	if av, existing := am.Get(akv.Key()); existing {
+		av.SetValue(akv.Value())
+	} else {
+		*am.orig = append(*am.orig, akv.orig)
+	}
+}
+
+// Len returns the number of AttributeKeyValue in the map.
+func (am AttributeMap) Len() int {
+	return len(*am.orig)
+}
+
+// Sort sorts the entries in the AttributeMap so two instances can be compared.
+// Returns the same instance to allow nicer code like:
+// assert.EqualValues(t, expected.Sort(), actual.Sort())
+func (am AttributeMap) Sort() AttributeMap {
+	sort.SliceStable(*am.orig, func(i, j int) bool { return (*am.orig)[i].Key < (*am.orig)[j].Key })
+	return am
+}
+
+// GetAttribute returns the AttributeKeyValue associated with the given index.
+//
+// This function is used mostly for itereting over all the values in the map:
+// for i := 1; i < am.Len(); i++ {
+//     akv := am.GetAttribute(i)
+//     ... // Do something with the attribute
+// }
+func (am AttributeMap) GetAttribute(ix int) AttributeKeyValue {
+	return AttributeKeyValue{(*am.orig)[ix]}
+}
+
 // AttributesMap stores a map of attribute keys to values.
+// TODO: Remove usage of this and use AttributeMap
 type AttributesMap map[string]AttributeValue
 
 // Attributes stores the map of attributes and a number of dropped attributes.
@@ -138,6 +292,7 @@ type Attributes struct {
 	droppedCount uint32
 }
 
+// NewAttributes creates a new Attributes with the given AttributesMap and droppedCount.
 func NewAttributes(m AttributesMap, droppedCount uint32) Attributes {
 	return Attributes{m, droppedCount}
 }