This PR plugs the new AttributeMap only into the Resource, but we can see a very nice benefit from this in the metrics benchmarks:

Before:
```
go test -benchmem -bench=. -run=notests ./internal/data/...
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector/internal/data
BenchmarkOtlpToFromInternal_PassThrough-16                              318627304                3.66 ns/op            0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_Int64Points_MutateOneLabel-16                2147724               551 ns/op             688 B/op         13 allocs/op
BenchmarkOtlpToFromInternal_DoublePoints_MutateOneLabel-16               2170123               554 ns/op             688 B/op         13 allocs/op
BenchmarkOtlpToFromInternal_HistogramPoints_MutateOneLabel-16            1902834               638 ns/op             768 B/op         14 allocs/op
BenchmarkOtlpToFromInternal_SummaryPoints_MutateOneLabel-16              1880725               639 ns/op             768 B/op         14 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector/internal/data 11.143s
```

After:
```
go test -benchmem -bench=. -run=notests ./internal/data/...
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector/internal/data
BenchmarkOtlpToFromInternal_PassThrough-16                              1000000000               0.256 ns/op           0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_Int64Points_MutateOneLabel-16                4681989               252 ns/op             240 B/op          7 allocs/op
BenchmarkOtlpToFromInternal_DoublePoints_MutateOneLabel-16               4736796               249 ns/op             240 B/op          7 allocs/op
BenchmarkOtlpToFromInternal_HistogramPoints_MutateOneLabel-16            3651998               331 ns/op             320 B/op          8 allocs/op
BenchmarkOtlpToFromInternal_SummaryPoints_MutateOneLabel-16              3555200               336 ns/op             320 B/op          8 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector/internal/data 7.523s
```