@bogdandrutu 2020-03-18
Based on https://github.com/open-telemetry/opentelemetry-collector/issues/653 I propose that:
1. First options is to have all the methods in the map to accept the immutable types
`InsertInt()`
`InsertBool()`
`InsertDouble()`
`InsertString()`
1. Second option is to make AttributeValue a real struct not backed by the proto (avoid unnecessary allocation that we cannot use it anyway). Then we can keep all the methods as they are.