@bogdandrutu 2020-03-18
@tigrannajaryan any comment about https://github.com/open-telemetry/opentelemetry-collector/pull/650#issuecomment-600811329?